Restructured point handling
